refactor fixes
